LibC: Reduce warnings in code

+++ userland/libc/dirent/readdir.c
@@ -1,4 +1,5 @@
 #include <dirent.h>
+#include <unistd.h>
 
 struct dirent *readdir(DIR *dir) {
   size_t offset = dir->dir_num * sizeof(struct dirent);
@@ -6,7 +7,7 @@ struct dirent *readdir(DIR *dir) {
   if (-1 == (rc = pread(dir->fd, &dir->internal_direntry, sizeof(struct dirent),
                         offset)))
     return NULL;
-  if (rc < sizeof(struct dirent))
+  if (rc < (int)sizeof(struct dirent))
     return NULL;
 
   dir->dir_num++;
